+++ resources/views/pages/partials/blocks/code.blade.php
@@ -16,11 +16,3 @@
 @endphp
 
 <pre>@if ($language !== '')<code data-language="{{ $language }}">@else<code>@endif{{ $block->content }}</code></pre>
-
-@if ($block->children->isNotEmpty())
-    <div class="wb-stack wb-gap-4">
-        @foreach ($block->children as $child)
-            @include('pages.partials.block', ['block' => $child])
-        @endforeach
-    </div>
-@endif