+++ resources/views/pages/partials/blocks/html.blade.php
@@ -13,11 +13,3 @@
 @endphp
 
 <div>{!! $trustedHtml['content'] ?? $block->content !!}</div>
-
-@if ($block->children->isNotEmpty())
-    <div class="wb-stack wb-gap-4">
-        @foreach ($block->children as $child)
-            @include('pages.partials.block', ['block' => $child])
-        @endforeach
-    </div>
-@endif