+++ resources/views/pages/partials/blocks/toc.blade.php
@@ -24,11 +24,3 @@
         </div>
     </div>
 @endif
-
-@if ($block->children->isNotEmpty())
-    <div class="wb-stack wb-gap-4">
-        @foreach ($block->children as $child)
-            @include('pages.partials.block', ['block' => $child])
-        @endforeach
-    </div>
-@endif